(Mac) Floor editor is now functional.
git-svn-id: svn://svn.code.sf.net/p/opensludge/code@39 27a91dd4-ec73-4ff0-838b-64dfa05f697e

+++ Mac Dev Kit/FloorDocument.m
@@ -44,6 +44,10 @@
     [super windowControllerDidLoadNib:aController];
 	[floorView connectToDoc: self];
 }
+- (void)mouseMoved:(NSEvent *)theEvent
+{
+	fprintf (stderr, ".");
+}
 
 - (BOOL)readFromURL:(NSURL *)absoluteURL 
 			 ofType:(NSString *)typeName 
@@ -83,6 +87,10 @@
 - (struct polyList *) getFloor {
 	return firstPoly;
 }
+- (void) setFloor:(struct polyList *) floor
+{
+	firstPoly = floor;
+}
 
 - (IBAction)loadBackdrop:(id)sender
 {
@@ -117,12 +125,24 @@
 
 - (void)close 
 {
-	forgetSpriteBank (&backdrop);
-	noFloor (&firstPoly);
-	deleteString (firstPoly);
+	if (backdrop.total) {
+		fprintf (stderr, "\n\n\nCLOSING!\n\n\n");
+		forgetSpriteBank (&backdrop);
+		backdrop.total = 0;
+		noFloor (&firstPoly);
+		deleteString (firstPoly);
+	}
 	[super close];
 }
 
+- (IBAction)changeMode:(id)sender
+{
+	mode = [modeSelector indexOfSelectedItem];
+}
+- (int) mode
+{
+	return mode;
+}
 
 @end
 
@@ -146,54 +166,209 @@
 - (void) setCoords {
 	glMatrixMode(GL_PROJECTION);
 	glLoadIdentity();
-	glOrtho(x+(-w/2)*zmul, x+(w/2)*zmul, y+(h/2)*zmul, y+(-h/2)*zmul, 1.0, -1.0);	
+	glOrtho(x*zmul, (x+w)*zmul, (-y+h)*zmul, -y*zmul, 1.0, -1.0);	
 	glMatrixMode(GL_MODELVIEW);	
+}
+
+- (BOOL)acceptsFirstResponder { return YES; }
+- (BOOL)resignFirstResponder{ return NO; }
+
+- (void)mouseMoved:(NSEvent *)theEvent
+{
+	NSPoint local_point = [self convertPoint:[theEvent locationInWindow] fromView:nil];
+	litX = (local_point.x+x)*zmul;
+	litY = -(local_point.y+y-h)*zmul;
+	if (lit != snapToClosest(&litX, &litY, [doc getFloor])){
+		lit = !lit;
+		[self setNeedsDisplay:YES];
+	}
 }
 
 - (void)mouseDown:(NSEvent *)theEvent {
     BOOL keepOn = YES;
+	int i, xx, yy;
     NSPoint mouseLoc;
 	NSPoint mouseLoc1;
+	NSPoint local_point;
 	
 	mouseLoc1 = [theEvent locationInWindow];
 
-	int x1 = x;
-	int y1 = y;
-	
-    while (keepOn) {
-        theEvent = [[self window] nextEventMatchingMask: NSLeftMouseUpMask |
-			NSLeftMouseDraggedMask];
-        mouseLoc = [theEvent locationInWindow];
+	unsigned int flags = [theEvent modifierFlags];
+	if (flags & NSCommandKeyMask || flags & NSControlKeyMask) {
+		local_point = [self convertPoint:mouseLoc1 fromView:nil];
+		xx = (local_point.x+x)*zmul;
+		yy = -(local_point.y+y-h)*zmul;
 		
-        switch ([theEvent type]) {
-            case NSLeftMouseUp:
-				keepOn = NO;
-				// continue
-            case NSLeftMouseDragged:
-				x = x1 + (mouseLoc1.x - mouseLoc.x)*zmul;
-				y = y1 + (mouseLoc.y - mouseLoc1.y)*zmul;
-				[self setCoords];
-				[self setNeedsDisplay:YES];
+		switch ([doc mode]) {
+		case 0: // Define floor border
+			snapToClosest(&xx, &yy, [doc getFloor]);
+			while (keepOn) {
+				keepOn = false;
+				i = addVertex (xx, yy, [doc getFloor]);
+				switch (i) {
+					case 0:
+						errorBox ("Can't add vertex", "Out of memory.");
+						return;
+						
+					case 3:
+						errorBox ("Can't add vertex", "That vertex is already used in this polygon, but isn't the start point.");
+						return;
+						
+					case 2:
+						if (NSRunAlertPanel (@"Can't add vertex", @"Can't add another vertex as the floor is already complete... do you want to start a NEW polygon at this point?", @"Yes", @"No", NULL) == NSAlertDefaultReturn) { 
+							[doc setFloor:addPoly ([doc getFloor])];
+							keepOn = true;
+						} else {
+							return;
+						}
+						break;
+				}
+			}
+			break;
+		case 1: // Move vertices
+			if (! snapToClosest(&xx, &yy, [doc getFloor]))
+				return;
+			
+			selx1 = xx;
+			sely1 = yy;
+			selection = 1;
+			while (keepOn) {
+				theEvent = [[self window] nextEventMatchingMask: NSLeftMouseUpMask | NSLeftMouseDraggedMask];
+				mouseLoc = [theEvent locationInWindow];
+				local_point = [self convertPoint:mouseLoc fromView:nil];
+				
+				switch ([theEvent type]) {
+					case NSLeftMouseUp:
+						keepOn = NO;
+						// continue
+					case NSLeftMouseDragged:
+						xx = (local_point.x+x)*zmul;
+						yy = -(local_point.y+y-h)*zmul;
+						
+						lit = snapToClosest(&xx, &yy, [doc getFloor]);
+						selx2 = litX = xx; 
+						sely2 = litY = yy;
+						
+						[self setNeedsDisplay:YES];
+						
+						break;
+					default:
+						/* Ignore any other kind of event. */
+						break;
+				}
+			}
+			selection = 0;
+			if (! moveVertices (selx1, sely1, selx2, sely2, [doc getFloor])) {
+				errorBox ("Can't move vertex", "Sorry - that vertex is already contained in one or more of the polygons you're changing.");
+				return;
+			}
+				
+			break;
+		case 2: // Remove vertices
+			if (! snapToClosest(&xx, &yy, [doc getFloor]))
+				return;
+			struct polyList * firstPoly = [doc getFloor];
+			
+			killVertex (xx, yy, &firstPoly);
+			[doc setFloor: firstPoly];
+		case 4: // Split lines
+		case 5: // Split segments
+			if (! snapToClosest(&xx, &yy, [doc getFloor]))
+				return;
+			
+			selx1 = xx;
+			sely1 = yy;
+			selection = 1;
+			while (keepOn) {
+				theEvent = [[self window] nextEventMatchingMask: NSLeftMouseUpMask | NSLeftMouseDraggedMask];
+				mouseLoc = [theEvent locationInWindow];
+				local_point = [self convertPoint:mouseLoc fromView:nil];
+				
+				switch ([theEvent type]) {
+					case NSLeftMouseUp:
+						keepOn = NO;
+						// continue
+					case NSLeftMouseDragged:
+						xx = (local_point.x+x)*zmul;
+						yy = -(local_point.y+y-h)*zmul;
+						
+						lit = snapToClosest(&xx, &yy, [doc getFloor]);
+						selx2 = litX = xx; 
+						sely2 = litY = yy;
+						
+						[self setNeedsDisplay:YES];
+						
+						break;
+					default:
+						/* Ignore any other kind of event. */
+						break;
+				}
+			}
+			selection = 0;
+			
+			if ([doc mode] == 4)
+				splitLine (selx1, sely1, selx2, sely2, [doc getFloor]);
+			else {
+				struct polyList * firstPoly = [doc getFloor];
+				splitPoly (selx1, sely1, selx2, sely2, &firstPoly);
+				[doc setFloor: firstPoly];
+			}
+			break;
+		}
+		lit = snapToClosest(&xx, &yy, [doc getFloor]);
+		litX = xx; litY = yy;
+		[self setNeedsDisplay:YES];
+		[doc updateChangeCount: NSChangeDone];	
+	} else {
+	
+		int x1 = x;
+		int y1 = y;
+		
+		while (keepOn) {
+			theEvent = [[self window] nextEventMatchingMask: NSLeftMouseUpMask |
+				NSLeftMouseDraggedMask];
+			mouseLoc = [theEvent locationInWindow];
+			
+			switch ([theEvent type]) {
+				case NSLeftMouseUp:
+					keepOn = NO;
+					// continue
+				case NSLeftMouseDragged:
+					x = x1 + (mouseLoc1.x - mouseLoc.x);
+					y = y1 - (mouseLoc.y - mouseLoc1.y);
+					[self setCoords];
+					[self setNeedsDisplay:YES];
 
-				break;
-            default:
-				/* Ignore any other kind of event. */
-				break;
-        }
-    };
+					break;
+				default:
+					/* Ignore any other kind of event. */
+					break;
+			}
+		}
+	}
 	
     return;
 }
 
 - (void)scrollWheel:(NSEvent *)theEvent
 {
+	if ([theEvent deltaY]) {
+		double x1, y1;
+		NSPoint mouseLoc1 = [theEvent locationInWindow];
+		NSPoint local_point = [self convertPoint:mouseLoc1 fromView:nil];
+		x1 = zmul*(local_point.x+x);
+		y1 = -zmul*(local_point.y+y-h);
+	
+		z += [theEvent deltaY];
+		if (z > 200.0) z = 200.0;
+		if (z < -16.0) z = -16.0;
+	
+		zmul = (1.0+z/20);
 
-	z += [theEvent deltaY];
-	if (z > 200.0) z = 200.0;
-	if (z < -16.0) z = -16.0;
-	
-	zmul = (1.0+z/20);
-	
+		x = -(local_point.x)+x1/zmul;
+		y = -(local_point.y-h)-y1/zmul;
+	}
+		
 	/*	
 	if ([theEvent deltaX]<0.0)
 		[self setValue:[NSNumber numberWithInt:spriteIndex+1] forKey:@"spriteIndex"];
@@ -214,12 +389,17 @@
 	r = g = b = 1.0;
 	zmul = (1.0+z/20);
 	[self setCoords];
+	[[self window] setAcceptsMouseMovedEvents:YES];
+	[[self window] makeFirstResponder: self];
+		
 }
 
 - (void)reshape {
 	NSRect bounds = [self bounds];
-	if (! w) x = bounds.size.width / 2;
-	if (! h) y = bounds.size.height / 2;
+	if (! w) x = 0;
+	else x -= (bounds.size.width-w)/2;
+	if (! h) y = 0;
+	else y += (bounds.size.height-h)/2;
 	h = bounds.size.height;
 	w = bounds.size.width;
 	glViewport (bounds.origin.x, bounds.origin.y, w, h);	
@@ -244,9 +424,31 @@
         glVertex3f(  0.0, 1000.0, 0.0);
     }
     glEnd();
+
+	if (lit) {
+		glColor3f(1.0f, 0.00f, 1.00f);
+		glBegin(GL_QUADS);
+		{
+			glVertex3i(  litX-8,  litY-8, 0);
+			glVertex3i(  litX+8,  litY-8, 0);
+			glVertex3i(  litX+8,  litY+8, 0);
+			glVertex3i(  litX-8,  litY+8, 0);
+		}
+		glEnd();
+	}
 	
 	glColor3f(r, g, b);	
 	drawFloor ([doc getFloor]);
+	
+	if (selection == 1) {
+		glColor3f(1.0f, 0.00f, 1.00f);
+		glBegin(GL_LINES);
+		{
+			glVertex3i(selx1, sely1, 0);
+			glVertex3i(selx2, sely2, 0);
+		}
+		glEnd();
+	}
 	
 	glFlush();
 }
